Move pki-related CLI utils into a separate module
