+++ lib/_cli/generate-csr.js
@@ -1,9 +1,10 @@
 'use strict'
 
+var pki = require('./pki')
 var util = require('./util')
 
 function generateCsr (configDir, commonName, command, verbosity) {
-  util.generatePrivateKeyAndCsr(configDir, commonName, command, verbosity)
+  pki.generatePrivateKeyAndCsr(configDir, commonName, command, verbosity)
 }
 
 module.exports = function (program) {
