+++ lib/_cli/provision-config.js
@@ -4,6 +4,7 @@ var fs = require('fs')
 var ini = require('ini')
 var path = require('path')
 var util = require('./util')
+var pki = require('./pki')
 
 var DEFAULT_BROKER_CERT_CHAIN_FILE_NAME = 'ca-bundle.crt'
 var PROVISION_COMMAND = 'DxlBrokerMgmt.generateOpenDXLClientProvisioningPackageCmd'
@@ -73,7 +74,7 @@ function storeProvisionConfig (config, configDir, filePrefix,
 
 function provisionConfig (configDir, hostName, commonOrCsrFileName, command,
                           verbosity) {
-  var results = util.generatePrivateKeyAndCsr(configDir, commonOrCsrFileName,
+  var results = pki.generatePrivateKeyAndCsr(configDir, commonOrCsrFileName,
     command, verbosity)
   var privateKeyFileName = results.privateKeyFileName
   var csrFileName = results.csrFileName
